Move Authentication Module
