+++ apps/api/src/app/app.module.ts
@@ -4,21 +4,25 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { MqttController } from './mqtt.controller';
 import { JwtModule } from '@nestjs/jwt';
-import { PassportController, PassportService, PassportModule as Wow } from '@event-participation-trends/passport';
+import {
+  PassportController,
+  PassportService,
+  PassportModule as Wow,
+} from '@event-participation-trends/api/passport';
 
 @Module({
   imports: [
     Wow,
     ClientsModule.register([
       {
-        name: "MQTT_SERVICE",
+        name: 'MQTT_SERVICE',
         transport: Transport.MQTT,
         options: {
           url: process.env.MQTT_URL,
           username: process.env.MQTT_USERNAME,
           password: process.env.MQTT_PASSWORD,
-        }
-      }
+        },
+      },
     ]),
     JwtModule.register({
       secret: process.env.JWT_SECRET,
