+++ libs/api/passport/README.md
@@ -1,0 +1,7 @@
+# api-passport
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test api-passport` to execute the unit tests via [Jest](https://jestjs.io).
