+++ libs/api/passport/project.json
@@ -1,21 +1,21 @@
 {
-  "name": "passport",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/passport/src",
+  "name": "api-passport",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/api/passport/src",
   "projectType": "library",
   "targets": {
     "lint": {
       "executor": "@nx/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/passport/**/*.ts"]
+        "lintFilePatterns": ["libs/api/passport/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/passport/jest.config.ts",
+        "jestConfig": "libs/api/passport/jest.config.ts",
         "passWithNoTests": true
       },
       "configurations": {
