+++ libs/api/passport/src/lib/google.strategy.ts
@@ -16,7 +16,7 @@ export class GoogleStrategy extends PassportStrategy(Strategy) {
     accessToken: string,
     refreshToken: string,
     profile: any,
-    done: VerifyCallback,
+    done: VerifyCallback
   ): Promise<any> {
     const { name, emails, photos } = profile;
     const user = {
