+++ libs/api/passport/src/lib/passport.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, UseGuards, Req, Get, Request, Res } from '@nestjs/common';
 import { Response as express_response } from 'express';
 import { PassportService } from './passport.service';
-import { GoogleOAuthGuard } from '../google-oauth-guard.guard'
+import { GoogleOAuthGuard } from '../google-oauth-guard.guard';
 
 @Controller('auth/')
 export class PassportController {
@@ -18,7 +18,7 @@ export class PassportController {
   async googleAuthRedirect(@Req() req: Request, @Res() res: express_response) {
     this.passportService.generateJWT(req).then((token) => {
       res.cookie('jwt', token, { httpOnly: true });
-      res.redirect('/api');
+      res.redirect(process.env['FRONTEND_URL'] || "");
     });
   }
 }
