+++ libs/app/login/feature/src/login.page.html
@@ -30,7 +30,7 @@
                 <ion-title ngClass="login-title">Welcome!</ion-title>
                 <ion-text color="none" ngClass="login-text">Log in/Sign up using a Google account</ion-text>
                 <ion-icon  ngClass="button-logo" name="logo-google" slot="start"></ion-icon>
-                <ion-button ngClass="login-button" color="none">
+                <ion-button href="http://localhost:3000/api/auth/google" ngClass="login-button" color="none">
                     <ion-label ngClass="button-text">Log in with Google</ion-label>
                 </ion-button>
                 <ion-text color="none" ngClass="login-text-tip">Once logged in/Signed up you will be able to view events.</ion-text>
