--- libs/passport/README.md
@@ -1,7 +1,0 @@
-# passport
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test passport` to execute the unit tests via [Jest](https://jestjs.io).
